+++ src/index.ts
@@ -4,4 +4,4 @@ export { stringToPath } from "./string-to-path";
 
 export type { Get } from "./get";
 export type { Set } from "./set";
-export type { PathString } from "./string-to-path";
+export type { StringToPath } from "./string-to-path";