+++ src/type-tests/get.test.types.ts
@@ -2,49 +2,63 @@ import { get, Get } from "../get";
 import { assert, Equals } from ".";
 
 // simple
-assert<Equals<Get<{ a: 5 }, ["a"]>, 5>>(true);
+assert<Equals<Get<{ a: 5 }, "a">, 5>>(true);
 
 // arrays
-assert<Equals<Get<[1, 2, 3], ["2"]>, 3>>(true);
+assert<Equals<Get<[1, 2, 3], "2">, 3>>(true);
 
 // mixed
-assert<
-  Equals<Get<{ a: [1, { b: ["a", "b"] }, 3] }, ["a", "1", "b", "1"]>, "b">
->(true);
+assert<Equals<Get<{ a: [1, { b: ["a", "b"] }, 3] }, "a.1.b.1">, "b">>(true);
 
 // wrong path
-assert<Equals<Get<{ a: 5 }, ["b"]>, undefined>>(true);
+assert<Equals<Get<{ a: 5 }, "b">, undefined>>(true);
 
 // default value
-assert<Equals<Get<{ a: 5 }, ["b"], "asdf">, "asdf">>(true);
+assert<Equals<Get<{ a: 5 }, "b", "asdf">, "asdf">>(true);
 
 // default value with undefined
-assert<Equals<Get<{ a: 5; b: undefined }, ["b"], "asdf">, "asdf">>(true);
+assert<Equals<Get<{ a: 5; b: undefined }, "b", "asdf">, "asdf">>(true);
 
 // default value with undefined in union
 assert<
-  Equals<Get<{ a: 5; b: number | undefined }, ["b"], "asdf">, number | "asdf">
+  Equals<Get<{ a: 5; b: number | undefined }, "b", "asdf">, number | "asdf">
 >(true);
 
 // should work with unions correctly
-assert<Equals<Get<{ a: number } | { a: string }, ["a"]>, string | number>>(
+// union objects
+assert<Equals<Get<{ a: number } | { a: string }, "a">, string | number>>(true);
+
+// union arrays
+assert<Equals<Get<{ a: { "2": "hello" } | [1, 2, 3] }, "a.2">, 3 | "hello">>(
   true
 );
 
+// union path
 assert<
-  Equals<Get<{ a: { "2": "hello" } | [1, 2, 3] }, ["a", "2"]>, 3 | "hello">
+  Equals<Get<{ a: number; b: { c: string } }, "a" | "b.c">, string | number>
+>(true);
+
+// union path and value
+assert<
+  Equals<
+    Get<
+      { a: number; b: { c: string } } | { a: symbol; b: { c: boolean } },
+      "a" | "b.c"
+    >,
+    string | number | symbol | boolean
+  >
 >(true);
 
 // should add undefined if one of union types is primitive
-assert<Equals<Get<{ a: number | [1, 2, 3] }, ["a", "2"]>, 3 | undefined>>(true);
+assert<Equals<Get<{ a: number | [1, 2, 3] }, "a.2">, 3 | undefined>>(true);
 
 // union type should add undefined for arrays (noUncheckIndexAccess)
-assert<Equals<Get<{ a: number[] }, ["a", "2"]>, number | undefined>>(true);
+assert<Equals<Get<{ a: number[] }, "a.2">, number | undefined>>(true);
 
 // regular arrays, primitive union
 assert<
   Equals<
-    Get<{ a: (number | { b: number })[] }, ["a", "2"]>,
+    Get<{ a: (number | { b: number })[] }, "a.2">,
     number | { b: number } | undefined
   >
 >(true);