+++ src/type-tests/string-to-path.test.types.ts
@@ -1,17 +1,17 @@
 import { assert, Equals } from "./index";
-import { PathString } from "../index";
+import { StringToPath } from "../index";
 
 // basic
-assert<Equals<PathString<"a.b.c.1.d">, ["a", "b", "c", "1", "d"]>>(true);
+assert<Equals<StringToPath<"a.b.c.1.d">, ["a", "b", "c", "1", "d"]>>(true);
 
 // empty
-assert<Equals<PathString<"">, []>>(true);
+assert<Equals<StringToPath<"">, []>>(true);
 
 // empty start
-assert<Equals<PathString<".a.b.1">, ["a", "b", "1"]>>(true);
+assert<Equals<StringToPath<".a.b.1">, ["a", "b", "1"]>>(true);
 
 // empty center
-assert<Equals<PathString<"a..b..1">, ["a", "b", "1"]>>(true);
+assert<Equals<StringToPath<"a..b..1">, ["a", "b", "1"]>>(true);
 
 // only dot
-assert<Equals<PathString<".">, []>>(true);
+assert<Equals<StringToPath<".">, []>>(true);