+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # CHANGELOG
 
+## 2.0.0
+
+## Added
+
+- `Get` and `Set` accept path as string instead of array
+
+## Fixed
+
+- `Get` and `Set` work correctly with union paths
+- `PathString` renamed to `StringToPath`
+
 ## 1.3.0
+
+### Fixed:
 
 - Get Performance and Recursion Depth Improvements ([#9](https://github.com/Ayub-Begimkulov/ts-get-set/pull/9))
 - Restructured a GetKey type to increase recursion depth for arrays ([#11](https://github.com/Ayub-Begimkulov/ts-get-set/pull/11))